+++ src/Loggable/Document/LogEntry.php
@@ -10,6 +10,7 @@
 namespace Gedmo\Loggable\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoODM;
+use Gedmo\Loggable\Document\Repository\LogEntryRepository;
 
 /**
  * Gedmo\Loggable\Document\LogEntry
@@ -24,6 +25,11 @@ use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoODM;
  *     }
  * )
  */
+#[MongoODM\Document(repositoryClass: LogEntryRepository::class)]
+#[MongoODM\Index(keys: ['objectId' => 'asc', 'objectClass' => 'asc', 'version' => 'asc'])]
+#[MongoODM\Index(keys: ['loggedAt' => 'asc'])]
+#[MongoODM\Index(keys: ['objectClass' => 'asc'])]
+#[MongoODM\Index(keys: ['username' => 'asc'])]
 class LogEntry extends MappedSuperclass\AbstractLogEntry
 {
     /*