+++ src/Translatable/Document/MappedSuperclass/AbstractTranslation.php
@@ -10,12 +10,14 @@
 namespace Gedmo\Translatable\Document\MappedSuperclass;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoODM;
+use Doctrine\ODM\MongoDB\Types\Type;
 
 /**
  * Gedmo\Translatable\Document\MappedSuperclass\AbstractTranslation
  *
  * @MongoODM\MappedSuperclass
  */
+#[MongoODM\MappedSuperclass]
 abstract class AbstractTranslation
 {
     /**
@@ -23,6 +25,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Id
      */
+    #[MongoODM\Id]
     protected $id;
 
     /**
@@ -30,6 +33,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Field(type="string")
      */
+    #[MongoODM\Field(type: Type::STRING)]
     protected $locale;
 
     /**
@@ -37,6 +41,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Field(type="string")
      */
+    #[MongoODM\Field(type: Type::STRING)]
     protected $objectClass;
 
     /**
@@ -44,6 +49,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Field(type="string")
      */
+    #[MongoODM\Field(type: Type::STRING)]
     protected $field;
 
     /**
@@ -51,6 +57,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Field(type="string", name="foreign_key")
      */
+    #[MongoODM\Field(name: 'foreign_key', type: Type::STRING)]
     protected $foreignKey;
 
     /**
@@ -58,6 +65,7 @@ abstract class AbstractTranslation
      *
      * @MongoODM\Field(type="string")
      */
+    #[MongoODM\Field(type: Type::STRING)]
     protected $content;
 
     /**