+++ tests/Gedmo/Translatable/Fixture/Document/Personal/Article.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Gedmo\Tests\Translatable\Fixture\Document\Personal;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoODM;
+use Doctrine\ODM\MongoDB\Types\Type;
 use Gedmo\Mapping\Annotation as Gedmo;
 use Gedmo\Tests\Translatable\Fixture\Personal\PersonalArticleTranslation;
 
@@ -19,20 +20,26 @@ use Gedmo\Tests\Translatable\Fixture\Personal\PersonalArticleTranslation;
  * @Gedmo\TranslationEntity(class="Gedmo\Tests\Translatable\Fixture\Document\Personal\ArticleTranslation")
  * @MongoODM\Document(collection="articles")
  */
+#[Gedmo\TranslationEntity(class: ArticleTranslation::class)]
+#[MongoODM\Document(collection: 'articles')]
 class Article
 {
     /** @MongoODM\Id */
+    #[MongoODM\Id]
     private $id;
 
     /**
      * @Gedmo\Translatable
      * @MongoODM\Field(type="string")
      */
+    #[Gedmo\Translatable]
+    #[MongoODM\Field(type: Type::STRING)]
     private $title;
 
     /**
      * @MongoODM\ReferenceMany(targetDocument="Gedmo\Tests\Translatable\Fixture\Document\Personal\ArticleTranslation", mappedBy="object")
      */
+    #[MongoODM\ReferenceMany(targetDocument: ArticleTranslation::class, mappedBy: 'object')]
     private $translations;
 
     /**